# frontend/openapi3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Booking and appointment management API.",
        "title": "Fejd API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/business/{businessID}/images": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads a business hero/logo/background image (multipart file). Visibility is public.",
                "tags": [
                    "admin"
                ],
                "summary": "Upload a business image",
                "parameters": [
                    {
                        "description": "Business UUID",
                        "name": "businessID",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "purpose": {
                                        "description": "hero|logo|background (default hero)",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/dto.Image"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/business/{slug}": {
            "get": {
                "description": "Returns a business with its services and employees by slug.",
                "tags": [
                    "public"
                ],
                "summary": "Get business details",
                "parameters": [
                    {
                        "description": "Business slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.BusinessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/business/{slug}/employees": {
            "get": {
                "description": "Returns all employees for a business.",
                "tags": [
                    "public"
                ],
                "summary": "List business employees",
                "parameters": [
                    {
                        "description": "Business slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/dto.BusinessUser"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/business/{slug}/sections": {
            "get": {
                "description": "Returns the salon's landing page sections ordered by position.",
                "tags": [
                    "public"
                ],
                "summary": "List landing page sections",
                "parameters": [
                    {
                        "description": "Business slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/dto.Section"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/business/{slug}/services": {
            "get": {
                "description": "Returns all active services for a business.",
                "tags": [
                    "public"
                ],
                "summary": "List business services",
                "parameters": [
                    {
                        "description": "Business slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/dto.Service"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/business/{slug}/slots": {
            "get": {
                "description": "Returns available time slots for a service and employee on a given date.",
                "tags": [
                    "public"
                ],
                "summary": "Get available time slots",
                "parameters": [
                    {
                        "description": "Business slug",
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Service UUID",
                        "name": "service_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Employee UUID",
                        "name": "employee_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Date (YYYY-MM-DD)",
                        "name": "date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.SlotsResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/handler.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "dto.Business": {
                "type": "object",
                "required": [
                    "created_at",
                    "id",
                    "name",
                    "slug",
                    "updated_at"
                ],
                "properties": {
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "dto.BusinessUser": {
                "type": "object",
                "required": [
                    "active",
                    "business_id",
                    "id",
                    "role",
                    "user_id"
                ],
                "properties": {
                    "active": {
                        "type": "boolean"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "business_id": {
                        "type": "string"
                    },
                    "display_name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                }
            },
...
            "dto.Image": {
                "type": "object",
                "required": [
                    "created_at",
                    "id",
                    "url"
                ],
                "properties": {
                    "content_type": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    }
                }
            },
...
            "dto.Section": {
                "type": "object",
                "required": [
                    "content",
                    "id",
                    "page_id",
                    "position",
                    "type"
                ],
                "properties": {
                    "content": {
                        "description": "Content is a locale-keyed JSON object.",
                        "type": "object"
                    },
                    "id": {
                        "type": "string"
                    },
                    "page_id": {
                        "type": "string"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "dto.Service": {
                "type": "object",
                "required": [
                    "active",
                    "business_id",
                    "created_at",
                    "duration_minutes",
                    "id",
                    "name"
                ],
                "properties": {
                    "active": {
                        "type": "boolean"
                    },
                    "business_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "duration_minutes": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "picture_id": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    }
                }
            },
            "dto.TimeSlot": {
                "type": "object",
                "required": [
                    "end_time",
                    "start_time"
                ],
                "properties": {
                    "end_time": {
                        "type": "string"
                    },
                    "start_time": {
                        "type": "string"
                    }
                }
            },
...
            "handler.BusinessImages": {
                "type": "object",
                "properties": {
                    "background": {
                        "type": "string"
                    },
                    "hero": {
                        "type": "string"
                    },
                    "logo": {
                        "type": "string"
                    }
                }
            },
            "handler.BusinessResponse": {
                "type": "object",
                "required": [
                    "business",
                    "employees",
                    "images",
                    "services"
                ],
                "properties": {
                    "business": {
                        "$ref": "#/components/schemas/dto.Business"
                    },
                    "employees": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/dto.BusinessUser"
                        }
                    },
                    "images": {
                        "$ref": "#/components/schemas/handler.BusinessImages"
                    },
                    "services": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/dto.Service"
                        }
                    }
                }
            },
...
            "handler.ErrorResponse": {
                "type": "object",
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "error message"
                    }
                }
            },
...
            "handler.SlotsResponse": {
                "type": "object",
                "required": [
                    "date",
                    "slots"
                ],
                "properties": {
                    "date": {
                        "type": "string",
                        "example": "2024-01-01"
                    },
                    "slots": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/dto.TimeSlot"
                        }
                    }
                }
            },
...
        }
    }
}